Add cast to argument in isdigit use
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@3831 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/libobjc/config/alpha/mframe-generic.h
@@ -81,7 +81,7 @@ typedef struct alpha_args {
     { \
       (TYPE)++; \
     } \
-  while (isdigit(*(TYPE))) \
+  while (isdigit ((int) *(TYPE))) \
     { \
       (TYPE)++; \
     } \
